image caption project/requirements.txt
fastapi 
uvicorn 
transformers
Pillow 
torch 
streamlit 
requests
python-multipart


